try fix

+++ exporter/exporter.go
@@ -104,7 +104,7 @@ func (ex *Exporter) Collect(ch chan<- prometheus.Metric) {
 				registered = 1
 			}
 			ch <- prometheus.MustNewConstMetric(trunkRegisteredDesc, prometheus.GaugeValue, float64(registered), labels...)
-			ch <- prometheus.MustNewConstMetric(trunkActiveCallsDesc, prometheus.GaugeValue, float64(trunk.SimCalls), labels...)
+			ch <- prometheus.MustNewConstMetric(trunkActiveCallsDesc, prometheus.GaugeValue, (trunk.SimCalls), labels...)
 		}
 	} else {
 		log.Println("failed to fetch TrunkList:", err)
